Keep the changes minimal by not moving the methods from Input task. Keep the input scripting evaluation lazy by introducing another closure in the middle.
